rest_api/loan_app/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
django-cors-headers>=4.7,<5.0
gunicorn>=23.0,<24.0
